+++ tests/Common/ManagerLogicTestTrait.php
@@ -15,15 +15,14 @@ use Yiisoft\Rbac\Exception\ItemAlreadyExistsException;
 use Yiisoft\Rbac\Exception\RuleNotFoundException;
 use Yiisoft\Rbac\Permission;
 use Yiisoft\Rbac\Role;
-use Yiisoft\Rbac\SimpleRuleFactory;
 use Yiisoft\Rbac\Tests\Support\AdsRule;
 use Yiisoft\Rbac\Tests\Support\AuthorRule;
 use Yiisoft\Rbac\Tests\Support\BanRule;
-use Yiisoft\Rbac\Tests\Support\EasyRule;
 use Yiisoft\Rbac\Tests\Support\FakeAssignmentsStorage;
 use Yiisoft\Rbac\Tests\Support\FakeItemsStorage;
 use Yiisoft\Rbac\Tests\Support\GuestRule;
 use Yiisoft\Rbac\Tests\Support\SubscriptionRule;
+use Yiisoft\Rbac\Tests\Support\TrueRule;
 
 trait ManagerLogicTestTrait
 {
@@ -175,30 +174,23 @@ trait ManagerLogicTestTrait
             ->createManager(
                 $this->createItemsStorage(),
                 $this->createAssignmentsStorage(),
-                new SimpleRuleFactory([
-                    'subscription' => new SubscriptionRule(),
-                    'ads' => new AdsRule(),
-                    'author' => new AuthorRule(),
-                    'ban' => new BanRule(),
-                    'guest' => new GuestRule(),
-                ]),
                 enableDirectPermissions: true,
             )
-            ->addRole((new Role('guest'))->withRuleName('guest'))
+            ->addRole((new Role('guest'))->withRuleName(GuestRule::class))
             ->setGuestRoleName('guest')
             ->addRole(new Role('news comment manager'))
             ->addRole(new Role('warned user'))
             ->addRole(new Role('trial user'))
-            ->addRole((new Role('subscribed user'))->withRuleName('subscription'))
-            ->addPermission((new Permission('view ads'))->withRuleName('ads'))
-            ->addPermission((new Permission('view ban warning'))->withRuleName('ban'))
+            ->addRole((new Role('subscribed user'))->withRuleName(SubscriptionRule::class))
+            ->addPermission((new Permission('view ads'))->withRuleName(AdsRule::class))
+            ->addPermission((new Permission('view ban warning'))->withRuleName(BanRule::class))
             ->addPermission(new Permission('view content'))
             ->addPermission(new Permission('view regular content'))
             ->addPermission(new Permission('view news'))
             ->addPermission(new Permission('add news comment'))
             ->addPermission(new Permission('view news comment'))
-            ->addPermission((new Permission('edit news comment'))->withRuleName('author'))
-            ->addPermission((new Permission('remove news comment'))->withRuleName('author'))
+            ->addPermission((new Permission('edit news comment'))->withRuleName(AuthorRule::class))
+            ->addPermission((new Permission('remove news comment'))->withRuleName(AuthorRule::class))
             ->addPermission(new Permission('view wiki'))
             ->addPermission(new Permission('view exclusive content'))
             ->addChild('view content', 'view regular content')
@@ -601,11 +593,11 @@ trait ManagerLogicTestTrait
     {
         $manager = $this->createFilledManager();
 
-        $rule = new EasyRule();
+        $rule = new TrueRule();
 
         $role = (new Role('new role'))
             ->withDescription('new role description')
-            ->withRuleName($rule->getName())
+            ->withRuleName(TrueRule::class)
             ->withCreatedAt(1_642_026_147)
             ->withUpdatedAt(1_642_026_148);
 
@@ -621,7 +613,7 @@ trait ManagerLogicTestTrait
             [
                 'name' => 'new role',
                 'description' => 'new role description',
-                'rule_name' => EasyRule::class,
+                'rule_name' => TrueRule::class,
                 'type' => 'role',
                 'updated_at' => 1_642_026_148,
                 'created_at' => 1_642_026_147,