+++ CHANGELOG.md
@@ -56,8 +56,8 @@
 - Bug #237: Handle not found base item in access tree (@arogachev)
 - Enh #245: Handle same names during renaming item in `AssignmentsStorage` (@arogachev)
 - Chg #208: Rename `getAccessTree()` to `getHierarchy()` in `ItemsStorageInterface` (@arogachev)
-- Enh #248: Add `SimpleRuleFactory` (@arogachev)
 - Enh #252: Return `$this` instead of throwing "already assigned" exception in `Manager::assign()` (@arogachev)
+- Enh #248: Add `SimpleRuleFactory` (@arogachev)
 
 ## 1.0.2 April 20, 2023